Minimize the duplication between DOT and String.
